requirements.txt
# We remove the Pillow pin and let the newer Streamlit handle it
fastapi==0.111.0
uvicorn==0.30.1
groq==0.9.0
faiss-cpu>=1.13.0
sentence-transformers==3.0.1
PyPDF2==3.0.1
streamlit>=1.40.0
python-dotenv==1.0.1
pydantic>=2.10.0
python-multipart==0.0.9
httpx==0.27.0